requirements.txt
pymupdf             
pillow              
paddleocr>=2.7.0  
requests
python-dotenv       
paddlepaddle
setuptools
transformers
sentencepiece
requests
torch
groq